Enhancements to Tailwind Class Generation and Figma Plugin Updates (#131)
* Added Tailwind class names for serif and mono fonts

When the user select a specific font family in Figma and it matches with the default font family defined in Tailwind, the generated code in the plugin uses the Tailwind utility classes.
https://tailwindcss.com/docs/font-family

* Added Tailwind class for overflow hidden

When the option "Clip content" in Figma on a frame, or component is enabled, the plugin prints the class "overflow-hidden" in the generated code.
https://tailwindcss.com/docs/overflow
https://www.delasign.com/blog/figma-frame-clip-content/

* Improvement to css class name generation for custom color names

Up until now the plugin would generate class names for custom color names as "hello-hello world" if the Figma variable name is "hello/hello world". This created 2 css class names.

With this code change spaces would be converted to dashes, so that the Figma variable name "hello/hello world" will generate "hello-hello-world".

* Added Figma "documentonchange" listener

Every time something is being updated in the Figma file, the generated plugin code updates.